Initial impl. for shrinking
